Refactored serializers into a new file
