+++ app/api/companies.py
@@ -9,10 +9,12 @@ DELETE  /v1/companies/:id   Deletes company with specific company_id
 
 """
 import os
+
 from app import db, cache
 from app.api import bp
 from app.errors.handlers import bad_request, error_response
-from app.models import Companies, Cities, Meta, companies_meta, CompaniesValidationSchema, CompaniesPatchSchema
+from app.models import Companies, Cities, Meta, companies_meta
+from app.serializers import CompaniesValidationSchema, CompaniesPatchSchema
 from app.utils_import import insert_meta, insert_city
 from app.utils import upload_file_to_s3, validate_image
 from config import Config
@@ -199,7 +201,7 @@ def delete_company(company_id):
     message = {}
     message["message"] = f"Company with company_id={company_id} has been deleted"
     response = jsonify(message)
-    response.status_code = 200
+    response.status_code = 204
 
     cache.clear()
     return response
